Separate critpath and non-critpath min_karma settings
FESCo has decided to reinstate different minimum karma for
critpath and non-critpath:
https://pagure.io/fesco/issue/3289
so we need to support this. This introduces critpath.min_karma,
and makes min_karma work much like mandatory_days_in_testing,
with the appropriate value for each update worked out at the
level of the update based on whether it's critpath or not.

We leave the default value of min_karma at 2 even though 1 would
be closer to the policy now, just because the tests were written
around 2 and are hard to adjust, especially the
test_autopush_reached tests in test_updates.py. We'll set it to
1 in downstream config.

Signed-off-by: Adam Williamson <awilliam@redhat.com>