bcd: introduce 'clean' subcommand for complete refresh
This adds a new subcommand to bcd which you can use to completely
refresh the environment. It removes all existing containers,
removes the bodhi-dev-bodhi and bodhi-dev-ipsilon images (so they
will be rebuilt on the next run), and pulls the latest versions
of all the base images we use (so all the containers will be
rebased on fresh versions of their base images at the next run).
This should give you a completely clean and updated bcd
environment the next time you do `./bcd run`.

Signed-off-by: Adam Williamson <awilliam@redhat.com>